+++ CHANGELOG.md
@@ -5,7 +5,7 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.14.0] - 2023-06-20
 
 ### Added
 
@@ -242,7 +242,7 @@ across our system.
 
 - An initial version.
 
-[Unreleased]: https://github.com/petabi/review-web/compare/0.13.1...main
+[0.14.0]: https://github.com/petabi/review-web/compare/0.13.1...0.14.0
 [0.13.1]: https://github.com/petabi/review-web/compare/0.12.0...0.13.1
 [0.13.0]: https://github.com/petabi/review-web/compare/0.12.0...0.13.0
 [0.12.0]: https://github.com/petabi/review-web/compare/0.11.0...0.12.0